+++ common/src/main/java/es/urjc/etsii/grafo/io/InstanceImporter.java
@@ -2,11 +2,12 @@ package es.urjc.etsii.grafo.io;
 
 import es.urjc.etsii.grafo.annotations.InheritedComponent;
 import es.urjc.etsii.grafo.exception.InstanceImportException;
+import es.urjc.etsii.grafo.util.IOUtil;
+import org.apache.commons.io.FilenameUtils;
 
 import java.io.BufferedReader;
-import java.io.File;
 import java.io.IOException;
-import java.nio.file.Files;
+import java.io.InputStreamReader;
 
 
 /**
@@ -21,14 +22,15 @@ public abstract class InstanceImporter<I extends Instance> {
      * Create an instance from the format used by the problem.
      * After this method finishes executing, the Instance object MUST BE IMMUTABLE.
      *
-     * @param f File from where we will load the data
+     * @param path File path from where we will load the data
      * @return The instance object that represents this object
      */
-    public I importInstance(File f){
-        try (var reader = Files.newBufferedReader(f.toPath())) {
-            return importInstance(reader, f.getName());
+    public I importInstance(String path){
+        try (var reader = new BufferedReader(new InputStreamReader(IOUtil.getInputStream(path)))) {
+            var suggestedInstanceId = FilenameUtils.getBaseName(path);
+            return importInstance(reader, suggestedInstanceId);
         } catch (Exception e) {
-            throw new InstanceImportException("Error while loading file: " + f.getAbsolutePath(), e);
+            throw new InstanceImportException("Error while loading file: " + path, e);
         }
     }
 
@@ -37,9 +39,9 @@ public abstract class InstanceImporter<I extends Instance> {
      * After this method finishes executing, the Instance object MUST BE IMMUTABLE.
      *
      * @param reader Input buffer, managed by the framework
-     * @param filename Filename on disk
+     * @param suggestedInstanceId Filename on disk
      * @throws IOException If an error is encountered while the instance is being parsed
      * @return The instance object that represents this object
      */
-    public abstract I importInstance(BufferedReader reader, String filename) throws IOException;
+    public abstract I importInstance(BufferedReader reader, String suggestedInstanceId) throws IOException;
 }