+++ common/src/main/java/es/urjc/etsii/grafo/util/IOUtil.java
@@ -1,6 +1,8 @@
 package es.urjc.etsii.grafo.util;
 
 
+import org.apache.commons.io.FilenameUtils;
+
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
@@ -12,12 +14,21 @@ import java.util.List;
 import java.util.Map;
 import java.util.logging.Logger;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
+import static org.apache.commons.compress.archivers.ArchiveStreamFactory.SEVEN_Z;
+import static org.apache.commons.compress.archivers.ArchiveStreamFactory.ZIP;
 
 /**
  * Util methods for managing input/output
  */
 public class IOUtil {
     private static final Logger log = Logger.getLogger(IOUtil.class.getName());
+
+    private static final Map<String, Compression.FileArchiveHandler> SUPPORTED_ARCHIVES = Map.of(
+            ZIP, new Compression.ZipArchiveHandler(),
+            SEVEN_Z, new Compression.SevenZHandler()
+    );
 
     /**
      * Create folder if not exists, recursively
@@ -35,9 +46,9 @@ public class IOUtil {
      *
      * @param path path to check
      */
-    public static void checkExists(String path){
+    public static void checkExists(String path) {
         File dir = new File(path);
-        if(!dir.exists()){
+        if (!dir.exists()) {
             throw new IllegalArgumentException("Path does not exist or not a folder: " + dir.getAbsolutePath());
         }
     }
@@ -47,9 +58,9 @@ public class IOUtil {
      *
      * @param path path to check
      */
-    public static void checkIsFolder(String path){
+    public static void checkIsFolder(String path) {
         File dir = new File(path);
-        if(!dir.isDirectory()){
+        if (!dir.isDirectory()) {
             throw new IllegalArgumentException("Path does not exist or not a folder: " + dir.getAbsolutePath());
         }
     }
@@ -60,7 +71,7 @@ public class IOUtil {
      * @param c class to check
      * @return true if inside a JAR, false otherwise
      */
-    public static boolean isJAR(Class<?> c){
+    public static boolean isJAR(Class<?> c) {
         String className = c.getName().replace('.', '/');
         String protocol = c.getResource("/" + className + ".class").getProtocol();
         return protocol.equals("jar");
@@ -69,13 +80,13 @@ public class IOUtil {
     /**
      * Get input stream for the given path
      *
-     * @param s path to resource
+     * @param s     path to resource
      * @param isJar true if the resource is inside a JAR file, false otherwise
      * @return InputStream to the resource given as a parameter
      * @throws java.io.IOException if anything goes wrong
      */
     public static InputStream getInputStreamForIrace(String s, boolean isJar) throws IOException {
-        if(isJar){
+        if (isJar) {
             return IOUtil.class.getResourceAsStream("/BOOT-INF/classes/irace/" + s);
             //return ResourceUtils.getFile("classpath:irace/" + s).toPath();
         } else {
@@ -86,9 +97,9 @@ public class IOUtil {
     /**
      * Get input stream for the given path
      *
-     * @param path path to resource
+     * @param path   path to resource
      * @param target where to copy the resource
-     * @param isJar true if the resource is inside a JAR file, false otherwise
+     * @param isJar  true if the resource is inside a JAR file, false otherwise
      * @throws java.io.IOException if anything goes wrong
      */
     public static void extractResource(String path, String target, boolean isJar, boolean autodelete) throws IOException {
@@ -99,11 +110,10 @@ public class IOUtil {
             pTarget = Path.of(target);
             Files.write(pTarget, inStream.readAllBytes());
         }
-        if(autodelete){
+        if (autodelete) {
             pTarget.toFile().deleteOnExit();
         }
     }
-
 
 
     /**
@@ -111,25 +121,25 @@ public class IOUtil {
      *
      * @param s path to file as string
      */
-    public static void markAsExecutable(String s){
+    public static void markAsExecutable(String s) {
         var f = new File(s);
         var success = f.setExecutable(true);
-        if(!success){
-            log.warning("Failed marking file as executable: "+s);
+        if (!success) {
+            log.warning("Failed marking file as executable: " + s);
         }
     }
 
     /**
      * Replace substitutions in the given input stream
      *
-     * @param origin input stream where the data is read from
-     * @param target where should data be written to
+     * @param origin        input stream where the data is read from
+     * @param target        where should data be written to
      * @param substitutions list of substitutions to do while copying the adta
      * @throws java.io.IOException if anything goes wrong
      */
     public static void copyWithSubstitutions(InputStream origin, Path target, Map<String, String> substitutions) throws IOException {
         String content = new String(origin.readAllBytes(), StandardCharsets.UTF_8);
-        for(var e: substitutions.entrySet()){
+        for (var e : substitutions.entrySet()) {
             content = content.replace(e.getKey(), e.getValue());
         }
         Files.writeString(target, content);
@@ -137,15 +147,16 @@ public class IOUtil {
 
     /**
      * Returns true if it is a file (not a folder) and the file is not hidden and its name does not start with a dot.
+     *
      * @param p Path to check
      * @return true if passes all checks, false otherwise
      */
-    public static boolean isNormalFile(Path p){
+    public static boolean isNormalFile(Path p) {
         try {
-            if(Files.isHidden(p)) {
+            if (Files.isHidden(p)) {
                 return false;
             }
-            if(p.toFile().getName().startsWith(".")){
+            if (p.toFile().getName().startsWith(".")) {
                 // In Windows files starting with . are not considered hidden, but they are in Linux, so ignore them.
                 return false;
             }
@@ -158,25 +169,73 @@ public class IOUtil {
 
     /**
      * List all normal files under a given path. Ignores hidden files. Recursively explores folders.
+     *
      * @param path Path to iterate
      * @return list of paths to normal files
      */
-    public static List<Path> iterate(Path path) {
-        try(var stream = Files.walk(path)) {
+    public static List<String> iterate(Path path) {
+        try (var stream = Files.walk(path)) {
             return stream
                     .filter(IOUtil::isNormalFile)
+                    .map(Path::toAbsolutePath)
+                    .flatMap(IOUtil::expandIfContainer)
                     .collect(Collectors.toList());
-        } catch (IOException e){
+        } catch (IOException e) {
             throw new RuntimeException(e);
         }
     }
 
     /**
      * List all normal files under a given path. Ignores hidden files. Recursively explores folders.
+     *
      * @param path Path to iterate
      * @return list of paths to normal files
      */
-    public static List<Path> iterate(String path){
+    public static List<String> iterate(String path) {
         return iterate(Path.of(path));
+    }
+
+    /**
+     * Check if the file represented by the given path is a compressed file, and if so, return list of files inside the archive.
+     *
+     * @param p Path to check
+     * @return list of files inside the archive, or the path itself if it is not a compressed file
+     */
+    public static Stream<String> expandIfContainer(Path p) {
+        String path = p.toAbsolutePath().toString();
+        if (path.contains(Compression.SEP)) {
+            throw new IllegalArgumentException("Path cannot not contain the compressed separator %s: %s".formatted(Compression.SEP, path));
+        }
+
+        if (Files.isDirectory(p)) {
+            throw new IllegalArgumentException("expandIfContainer should only be called with files, not directories");
+        }
+
+        var fileExtension = FilenameUtils.getExtension(path);
+        var handler = SUPPORTED_ARCHIVES.get(fileExtension);
+        if (handler == null) {
+            // Not a known compressed filetype, return file path as is
+            return Stream.of(p.toString());
+        }
+        // else, it is a compressed file we know how to handle, return list of files inside the archive
+        return handler.getFileEntries(path);
+    }
+
+    public static InputStream getInputStream(String path) throws IOException {
+        if (!path.contains(Compression.SEP)) {
+            // Not a compressed file, just return the input stream
+            return Files.newInputStream(Path.of(path));
+        }
+        // Compressed file, find the entry and return the input stream
+        var split = path.split(Compression.SEP);
+        var archivePath = split[0];
+        var entryPath = split[1];
+        var fileExtension = FilenameUtils.getExtension(archivePath);
+        var archiver = SUPPORTED_ARCHIVES.get(fileExtension);
+        if (archiver == null) {
+            throw new IllegalArgumentException("Path is compressed, but the file extension is not recognized: " + archivePath);
+        }
+
+        return archiver.getEntryInputStream(archivePath, entryPath);
     }
 }